ARTS-95 Clean up and fix auth init process
