+++ src/app/app.component.ts
@@ -20,23 +20,9 @@ export class AppComponent {
               public apiEndopintService: ApiEndpointService) {
     this.configurationService.init(appRoot);
     this.apiEndopintService.init(this.configurationService.gatewayUrl);
-    this.authenticationService.init();
+    this.authenticationService.init(appRoot);
     console.log('AppComponent constructor config', configurationService);
     console.log('AppComponent constructor profileService', this.apiEndopintService.profileEndpoint);
-  }
-
-  private initAuthenticationConfig(appRoot: ElementRef<any>): void {
-    const issuer = appRoot.nativeElement.getAttribute('issuer');
-    const clientId = appRoot.nativeElement.getAttribute('clientId');
-
-    //  We may be passed some config through the app-root element
-    if (issuer != null && issuer.length > 0 && issuer.indexOf('$') !== 0) {
-      console.log(`Initing auth service with app-root config - issuer: ${authConfig.issuer} client: ${authConfig.clientId}`);
-      authConfig.issuer = issuer;
-      authConfig.clientId = clientId;
-    } else {
-      console.log(`Initing auth service from angular env - issuer: ${environment.issuer} client: ${environment.clientId}`);
-    }
   }
 
   logout(): void {
